+++ storybook/src/Theme/Font.stories.tsx
@@ -24,9 +24,55 @@
 
  */
 
-import { FontFamilyChoices } from '../../system'
+import {
+  ComponentsProvider,
+  FieldText,
+  Fieldset,
+  SpaceVertical,
+  Heading,
+  Paragraph,
+  theme,
+} from '@looker/components'
+import React, { FC, FormEvent, useState } from 'react'
 
-export const fontFamilies: FontFamilyChoices = {
-  brand: `'Open Sans', 'Noto Sans JP', 'Noto Sans CJK KR', 'Noto Sans Arabic UI', 'Noto Sans Devanagari UI', 'Noto Sans Hebrew', 'Noto Sans Thai UI', Helvetica, Arial, sans-serif, 'Noto Sans'`,
-  code: `'Monaco', 'Menlo', 'Ubuntu Mono', 'Consolas', 'source-code-pro', monospace`,
+export const Font: FC = () => {
+  const [brand, setBrand] = useState('Comic Sans MS')
+
+  const handleBrandChange = (event: FormEvent<HTMLInputElement>) => {
+    setBrand(event.currentTarget.value)
+  }
+
+  const actualTheme = {
+    ...theme,
+    fonts: {
+      ...theme.fonts,
+      brand,
+    },
+  }
+
+  // const coreColors = useState<CoreColors>(theme)
+
+  return (
+    <SpaceVertical>
+      <Fieldset legend={name}>
+        <FieldText
+          label="Brand Font Face"
+          value={brand}
+          onChange={handleBrandChange}
+        />
+      </Fieldset>
+
+      <pre>{String(theme)}</pre>
+
+      <ComponentsProvider theme={actualTheme}>
+        <Heading>A Header</Heading>
+        <Paragraph>Paragraph text.</Paragraph>
+      </ComponentsProvider>
+    </SpaceVertical>
+  )
+}
+
+export default {
+  component: Font,
+  title: 'Theme',
 }