+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- `ComponentsProvider` actually uses the `theme` passed in
 - `ButtonGroup` space between rows when wrapping
 - `InputChips` separates chips by newline when pasting